Restructure component paths under a unified 'selectors' folder
Moved FactionSelector, EnemySelector, and RaritySelector components into a dedicated 'selectors' subfolder. Updated imports across the app to reflect the new paths, ensuring a cleaner and more consistent project structure.
