+++ src/components/selectors/rarity-selector.tsx
